feat: actions sortable list
